generate timestamp fcn (separation)
